+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Removed `benf\neo\converters\models\BlockTypeGroup` (Neo BlockTypeGroup Converter for Schematic)
 - Removed `benf\neo\elements\Block::getModified()`; use `$dirty` instead
 - Removed `benf\neo\elements\Block::setModified()`; use `$dirty` instead
+- Removed `benf\neo\services\Blocks::getSearchKeywords()`
 - Removed `benf\neo\services\Fields::getSupportedSiteIdsForField()`
 
 ## 2.11.13 - 2021-09-15